Fix some bug and imple healState

+++ PlatformerProject/Assets/_Scripts/Core/CoreComponents/Combat.cs
@@ -26,7 +26,7 @@ public class Combat : CoreComponent, IDamageable, IKnockbackable
     {
         CheckKnockback();
     }
-    public void Damage(float amount)
+    public void Damage(int amount)
     {
         if (canDamage)
         {
