+++ PlatformerProject/Assets/_Scripts/Enemies/States/Data/D_MeleeAttack.cs
@@ -6,7 +6,7 @@ using UnityEngine;
 public class D_MeleeAttack : ScriptableObject
 {
     public float attackRadius = 0.5f;
-    public float attackDamage = 10f;
+    public int attackDamage = 10;
     public float knockbackStrength = 10f;
     public Vector2 knockbackAngle = Vector2.one;
     public LayerMask whatIsPlayer;
